+++ .gitignore
@@ -1,3 +1,4 @@
 joho-ta-tools
 .vscode
 argparse
+main.cpp
