+++ README.md
@@ -1,5 +1,5 @@
 # .joho-ta-tools
-testcase作成用のツール
+Online JudgeにおけるTAの仕事の効率化を目指します。 
 
 ## Environment
 - Ubuntu 18.04
@@ -16,23 +16,25 @@ echo "PATH=PATH:~/.joho-ta-tools" >> ~/.profile
 source ~/.profile
 ```
 
-## How to use
+## generate testcase
+testcase作成の自動化ツール。模範解答(.c/.cpp)は手元にあること前提でそこからテストケースを半自動作成します。randomモードが一番使い勝手がいいと思います。
+### How to use
 1. テストケースの模範解答(`main.cpp`)があるディレクトリに移動
-2. `joho-ta-tools`でディレクトリ内の`main.cpp`がコンパイルされ、入力待ちの状態になる。
+2. `joho-ta-tools gen`でディレクトリ内の`main.cpp`がコンパイルされ、入力待ちの状態になる。
 3. `Input?:`に続いて、テストケースの入力を行う。入力が終わったら、`Enter`を入力。
 4. `Output?:`として、計算結果が出力される。
 5. テストケースの作成を終えたいときは、`Input?:`の時に`Enter`を入力する。
 6. 生成物として、`*.in`, `*.out`, `testcase.zip`が生成される。
 
-## How to use with special judge
+### with special judge
 上記の1, 3, 4, 5は同じ。
 
-2'. `joho-ta-tools -s`でディレクトリ内の`main.cpp`がコンパイルされ、入力待ちの状態になる。
+2'. `joho-ta-tools gen -s`でディレクトリ内の`main.cpp`がコンパイルされ、入力待ちの状態になる。
 
 6'. `*.in`, `testcase.zip`, `spj.cpp`が出力される。
 
-## How to use with random mode
-1. `joho-ta-tools -r num_of_samples`を入力
+### with random mode
+1. `joho-ta-tools gen -r num_of_samples`を入力
 2. `Input format:`に続いて、入力の型を決定する。入力の際のルールは以下に従う。
    ```
    int : i min max
@@ -59,5 +61,5 @@ source ~/.profile
 
 
 ## 注意
-- random modeで、文字範囲の入力はASCIIコード表に準ずるため、A~zを指定した時、`[, \, ]`などの記号も含まれる。
+- random modeで、文字範囲の入力はASCIIコード表に準ずるため、A~zを指定した時、`[, \, ]`などの記号も含まれる。小文字のみか大文字のみを指定することを推奨。
 - 色々バグがあると思うんで、バグの再現例とissueかpull reqお願いします。
